### 2024-07-17 #7001 @kthyng
@hugoledoux Can you add your affiliation to the paper?
