### 2024-07-17 642fdc4
Make affiliation explicit even if one author only
