+++ paper/paper.md
@@ -9,8 +9,10 @@ tags:
 authors:
   - name: Hugo Ledoux
     orcid: 0000-0002-1251-8654
+    affiliation: "1"
 affiliations:
-  - name: Delft University of Technology, the Netherlands
+ - name: Delft University of Technology, the Netherlands
+   index: 1
 date: 17 July 2024
 bibliography: ref.bib
 
