### 2024-07-17 #7001 @hugoledoux
@kthyng okay fixed it. I was puzzled because the affiliation was there and I had successfully compiled it with the github action... but I wasn't using the latest v4 it seems, and things have changed. Mystery solved.
