### 2024-07-19 #7001 @kthyng
Hi @hugoledoux! Two things:
* can you add a short statement in your summary or statement of need to clarify ties to research?
* please shorten the paper. I'd suggest either keeping the sections but shortening them (some of this is repetitive) or remove the implementation details and example. We're trying to be better about keeping the papers shorter.
Thanks!
