### 2024-07-22 c69fecd
Work on making hte paper shorter
I put the removed material to the docs
