### 2024-07-22 6bd38e2
Add comparison of startinpy to the docs, extra page
