### 2024-07-22 3e56d9d
Fix typos in new version paper
