### 2024-07-22 df3f810
Copy new paper
