+++ paper/paper2.md
@@ -1,76 +1,0 @@
----
-title: 'startinpy: A Python library for modelling and processing 2.5D triangulated terrains'
-tags:
-  - Delaunay triangulation
-  - Python
-  - terrain
-  - interpolation
-  - GIS
-authors:
-  - name: Hugo Ledoux
-    orcid: 0000-0002-1251-8654
-affiliations:
- - name: Delft University of Technology, the Netherlands
-date: 17 July 2024
-bibliography: ref.bib
-
----
-
-# Summary
-
-*startinpy* is a Python library for modelling and processing terrains using a 2D Delaunay triangulation (DT).
-The triangulation is computed in 2D, but the z-elevation of the vertices is kept (which is referred to as 2.5D modelling).
-
-Given a dataset formed of elevation samples (eg collected with lidar or photogrammetry), it allows us, among other things, to reconstruct a terrain, to remove points, to search efficiently in the triangulation, to attach attributes to the vertices, and to convert to a gridded terrain (a few spatial interpolation methods based on the DT and/or its dual structure, the Voronoi diagram, have been implemented: linear interpolation, natural neighbours [@Sibson81], and Laplace interpolation [@Belikov97]).
-
-The core of the library is written in the language Rust (so it can process large datasets quickly), robust arithmetic is used for the updating of the DT (the robust predicates of @Shewchuk96 are used), and it uses NumPy for input/output of data, which allows it to integrate with other Python libraries used by researchers.
-
-The core of the library has been used recently to build a global coastal terrain using laser altimetric measurements from the space station [@Pronk24], and it has been used for a few projects dealing with aerial and space lidar datasets, eg @Meschin23 and @Kan24.
-
-While startinpy was developed primarily for terrains, it can be used as an easy-to-use and fast 2D Delaunay triangulator (and Voronoi diagram generator), which, as elaborated in @Aurenhammer12, are two structures that play an essential role in several disciplines: astronomy, geology, ecology, engineering, etc.
-
-
-# Statement of need
-
-While there exist many Python libraries for computing the DT in 2D (a search for "Delaunay triangulation" in the *Python Package Index* (PyPI) returns 227 packages), most of them are not fully suitable for the modelling of 2.5D triangulated terrain.
-
-startinpy has the following properties, which greatly improve the modelling and processing of 2.5D terrains.
-
-**It is fast for large datasets.**
-With a recent lidar scanner, we can easily collect 50 samples/$m^2$, which means that a 1$km^2$ area will contain 50+ million samples. 
-Since constructing a DT requires several steps, if those steps are implemented in pure Python then the library becomes very slow.
-As can be seen in the [DT construction comparison](https://startinpy.readthedocs.io/latest/comparison.html), startinpy is faster than most other libraries for large datasets.
-This is partly because it is 100% developed in Rust; the core library is called "startin" and its the source code is available at https://github.com/hugoledoux/startin.
-
-**Its data structure is exposed.**
-Most libraries only return a list of vertices and triangles (triplets of vertex identifiers), which means that the user has to build an auxiliary graph to be able to find the adjacent triangles of a given one, or to find all the triangles that are incident to a given vertex (eg to calculate the normal).
-The library's name comes from the fact that the data structure implemented is based on the concept of *stars* in a graph [@Blandford05], which allows us to store adjacency and incidence, and have a very compact data structure.
-startinpy exposes methods to search triangles, to find the adjacent triangles of a triangle and the incident triangles to a vertex.
-
-
-**The DT is incrementally constructed and deletion of vertices is possible.**
[…]
