### 2024-07-22 d381846
Add again polycope fig to paper
