### 2024-07-22 #7001 @hugoledoux
@kthyng I have cut the paper in half, and moved the material to the docs online: https://startinpy.readthedocs.io/latest/
