### 2024-08-12 #7001 @mikemahoney218
@weiji14 -- would you be willing to review a submission for the Journal of Open Source Software?

At JOSS, we carry out our checklist-driven reviews here in GitHub issues, encourage author-reviewer-editor conversations, and reviews involve downloading and installing the software, and inspecting the repository and submitted paper. See https://joss.readthedocs.io/en/latest/review_criteria.html for more about our review process.

Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review). We aim for the initial review to be completed within 4 weeks (which would bring us to September 9th), though we can of course be flexible as needed.

Would you be willing to review this submission for JOSS? If not, can you recommend any other potential reviewers?
