### 2024-08-12 #7001 @weiji14
Oo nice, a maturin project, and working with LiDAR data! Yes I would be more than happy to review this. I will note that I'll be travelling for two weeks until 1st September, so there might be delays in finding time to review, but I'll let you know if I need more time beyond the 9th September date.
