### 2024-08-20 #7001 @evetion
Given this [PR](https://github.com/openjournals/joss-reviews/issues/6937), maybe @DanielVandH can review?
