### 2024-08-20 #7123 @editorialbot
[REVIEW]: startinpy: A Python library for modelling and processing 2.5D triangulated terrains
**Submitting author:** @hugoledoux (<a href="http://orcid.org/0000-0002-1251-8654">Hugo Ledoux</a>)
**Repository:** https://github.com/hugoledoux/startinpy
**Branch with paper.md** (empty if default branch):
**Version:** 0.12.0
