### 2024-08-20 #7123 @mikemahoney218
👋🏼 @hugoledoux @weiji14 @kylemann16 this is the review thread for the paper. Just about of our communications will happen here from now on 😄

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4 weeks. Please let me know if you require some more time.

Please feel free to ping me (@mikemahoney218) if you have any questions/concerns. Thank you so much for reviewing for JOSS!
