### 2024-08-20 #7123 @mikemahoney218
👋 Pete, this is the review thread for startinpy. Hopefully my notes above make sense -- please feel free to reach out if you have any questions or anything comes up during your review. Thank you so much for reviewing for JOSS!
