### 2024-08-21 292e6eb
Fix the 2 DOIs in the references
