+++ paper/ref.bib
@@ -1,7 +1,7 @@
 
 @book{Aurenhammer12,
   author = {Aurenhammer, Franz and Klein, Rolf and Lee, Der-Tsai},
-  doi = {http://dx.doi.org/10.1142/8685},
+  doi = {10.1142/8685},
   publisher = {World Scientific},
   title = {Voronoi Diagrams and Delaunay Triangulations},
   year = {2012}}
@@ -11,6 +11,7 @@
   author = {Cl\'ement Mallet and Fr\'ed\'eric Bretar},
   date-modified = {2013.06.25},
   journal = {ISPRS Journal of Photogrammetry and Remote Sensing},
+  doi = {10.1016/j.isprsjprs.2008.09.007},
   number = {1},
   pages = {1--16},
   title = {Full-waveform topographic lidar: State-of-the-art},
