### 2024-08-21 fcaf467
docs: grammar, punctuation, and wording suggestions
These suggestions came out of an initial review of the JOSS paper as a part of
https://github.com/openjournals/joss-reviews/issues/7123. These are merely
suggestions, and by no means are required to be taken up during the review
process. I just thought I'd do a quick-once over while reading through.
