+++ paper/paper.md
@@ -20,55 +20,52 @@ bibliography: ref.bib
 
 # Summary
 
-*startinpy* is a Python library for modelling and processing terrains using a 2D Delaunay triangulation (DT).
+*startinpy* is a Python library for modelling and processing terrains using a two-dimensional (2D) Delaunay triangulation (DT).
 The triangulation is computed in 2D, but the z-elevation of the vertices is kept (which is referred to as 2.5D modelling).
 
-Given a dataset formed of elevation samples (eg collected with lidar or photogrammetry), it allows us, among other things, to reconstruct a terrain, to remove points, to search efficiently in the triangulation, to attach attributes to the vertices, and to convert to a gridded terrain (a few spatial interpolation methods based on the DT and/or its dual structure, the Voronoi diagram, have been implemented: linear interpolation, natural neighbours [@Sibson81], and Laplace interpolation [@Belikov97]).
+Given a dataset formed of elevation samples (e.g. collected with lidar or photogrammetry), startinpy allows us to reconstruct a terrain, to remove points, to search efficiently in the triangulation, to attach attributes to the vertices, and to convert to a gridded terrain
+Additional functionality is provided, including a few spatial interpolation methods based on the DT and/or its dual structure, the Voronoi diagram, have been implemented: linear interpolation, natural neighbours [@Sibson81], and Laplace interpolation [@Belikov97].
 
-The core of the library is written in the language Rust (so it can process large datasets quickly), robust arithmetic is used for the updating of the DT (the robust predicates of @Shewchuk96 are used), and it uses NumPy for input/output of data, which allows it to integrate with other Python libraries used by researchers.
+The core of the library is written in [Rust](https://www.rust-lang.org/) (so it can process large datasets quickly), robust arithmetic is used for the updating of the DT (the robust predicates of @Shewchuk96 are used), and it uses NumPy for input/output of data, which allows it to integrate with other Python libraries used by researchers.
 
-The core of the library has been used recently to build a global coastal terrain using laser altimetric measurements from the space station [@Pronk24], and it has been used for a few projects dealing with aerial and space lidar datasets, eg @Meschin23 and @Kan24.
+The core of the library has been used to build a global coastal terrain using laser altimetric measurements from the space station [@Pronk24], and it has been used for several projects dealing with aerial and space lidar datasets e.g. @Meschin23 and @Kan24.
 
-While startinpy was developed primarily for terrains, it can be used as an easy-to-use and fast 2D Delaunay triangulator (and Voronoi diagram generator), which, as elaborated in @Aurenhammer12, are two structures that play an essential role in several disciplines: astronomy, geology, ecology, engineering, etc.
+While startinpy was developed primarily to build terrains, it can be used as an easy-to-use and fast 2D Delaunay triangulator (and Voronoi diagram generator), which, as elaborated in @Aurenhammer12, are two structures that play an essential role in several disciplines: astronomy, geology, ecology, engineering, etc.
 
 ![A lidar dataset terrain reconstructed with startinpy and visualised with another Python library (Polyscope).](polyscope_gui.jpg)
 
 # Statement of need
 
-While there exist many Python libraries for computing the DT in 2D (a search for "Delaunay triangulation" in the *Python Package Index* (PyPI) returns 227 packages), most of them are not fully suitable for the modelling of 2.5D triangulated terrain.
+While there exist many Python libraries for computing the DT in 2D (a search for "Delaunay triangulation" in the [*Python Package Index* (PyPI)](https://pypi.org/search/?q=Delaunay+triangulation) returns over 200 packages), most of them are not fully suitable for the modelling of 2.5D triangulated terrain.
 
 startinpy has the following properties, which greatly improve the modelling and processing of 2.5D terrains.
 
 **It is fast for large datasets.**
[…]
