### 2024-08-21 756c44a
docs: add notes about running the test suite
Includes a little diff fluff from my Markdown auto-formatter, hope that's
alright.
