+++ tests/requirements.in
@@ -1,0 +1,3 @@
+laspy[lazrs]
+numpy
+pytest
