### 2024-08-22 #7123 @hugoledoux
btw @gadomski has already started and made 2 PR directly in the branch main of startinpy, which is fine. My plan is to work directly on main (code+paper) during this review and at the end bump bump the version (0.12 or 1.0, depending on what you guys suggest also).
