### 2024-08-22 5be1b98
Some corrections on the corrections of the paper
