+++ paper/paper.md
@@ -23,14 +23,14 @@ bibliography: ref.bib
 *startinpy* is a Python library for modelling and processing terrains using a two-dimensional (2D) Delaunay triangulation (DT).
 The triangulation is computed in 2D, but the z-elevation of the vertices is kept (which is referred to as 2.5D modelling).
 
-Given a dataset formed of elevation samples (e.g. collected with lidar or photogrammetry), startinpy allows us to reconstruct a terrain, to remove points, to search efficiently in the triangulation, to attach attributes to the vertices, and to convert to a gridded terrain
+Given a dataset formed of elevation samples (eg collected with lidar or photogrammetry), startinpy allows us to reconstruct a terrain, to remove points, to search efficiently in the triangulation, to attach attributes to the vertices, and to convert to a gridded terrain.
 Additional functionality is provided, including a few spatial interpolation methods based on the DT and/or its dual structure, the Voronoi diagram, have been implemented: linear interpolation, natural neighbours [@Sibson81], and Laplace interpolation [@Belikov97].
 
 The core of the library is written in [Rust](https://www.rust-lang.org/) (so it can process large datasets quickly), robust arithmetic is used for the updating of the DT (the robust predicates of @Shewchuk96 are used), and it uses NumPy for input/output of data, which allows it to integrate with other Python libraries used by researchers.
 
-The core of the library has been used to build a global coastal terrain using laser altimetric measurements from the space station [@Pronk24], and it has been used for several projects dealing with aerial and space lidar datasets e.g. @Meschin23 and @Kan24.
+The core of the library has been used to build a global coastal terrain using laser altimetric measurements from the space station [@Pronk24], and it has been used for several projects dealing with aerial and space lidar datasets, for instance @Meschin23 and @Kan24.
 
-While startinpy was developed primarily to build terrains, it can be used as an easy-to-use and fast 2D Delaunay triangulator (and Voronoi diagram generator), which, as elaborated in @Aurenhammer12, are two structures that play an essential role in several disciplines: astronomy, geology, ecology, engineering, etc.
+While startinpy was developed primarily to build and manipulate terrains, it can be used as an easy-to-use and fast 2D Delaunay triangulator (and Voronoi diagram generator), which, as elaborated in @Aurenhammer12, are two structures that play an essential role in several disciplines: astronomy, geology, ecology, engineering, etc.
 
 ![A lidar dataset terrain reconstructed with startinpy and visualised with another Python library (Polyscope).](polyscope_gui.jpg)
 
@@ -47,7 +47,7 @@ As can be seen in the [DT construction comparison](https://startinpy.readthedocs
 This is partly because it is 100% developed in Rust; the core library is called "startin" and its the source code is available at <https://github.com/hugoledoux/startin>.
 
 **Its data structure is exposed.**
-Most libraries only return a list of vertices and triangles (triplets of vertex identifiers), which means that the user has to build an auxiliary graph to be able to find the adjacent triangles of a given one, or to find all the triangles that are incident to a given vertex (e.g. to calculate the normal).
+Most libraries only return a list of vertices and triangles (triplets of vertex identifiers), which means that the user has to build an auxiliary graph to be able to find the adjacent triangles of a given one, or to find all the triangles that are incident to a given vertex (eg to calculate the normal).
 This library's name comes from the fact that the data structure implemented is based on the concept of *stars* in a graph [@Blandford05], which allows us to store adjacency and incidence, and have a very compact data structure.
 startinpy exposes methods to search triangles, find the adjacent triangles of a triangle, and find the incident triangles to a vertex.
 
@@ -58,12 +58,12 @@ The deletion of vertices in a DT is useful to remove outliers (which are detecte
 
 **The z-values are stored and xy-duplicates handled.**
[…]
