### 2024-08-22 hugoledoux/startinpy#23 @hobu
> `Filter.range(limits="Classification[2:2]")`

`Filters.expression(expression="Classification == 2")` <-- much better modern PDAL way to the same thing.
