### 2024-08-22 4f5ec53
chore: format comparisons.py
