+++ dt_comparisons/comparisons.py
@@ -1,23 +1,24 @@
-import numpy as np
-import laspy
-import rasterio
 import time
-import startinpy
-import triangle 
-from scipy.spatial import Delaunay
+from random import uniform
+
+import laspy
+import numpy as np
+import rasterio
+import triangle
+from Delaunator import Delaunator
+from delaunay.delaunay import delaunay
 from delaunay.quadedge.mesh import Mesh
 from delaunay.quadedge.point import Vertex
-from delaunay.delaunay import delaunay
-from random import uniform
-from Delaunator import Delaunator
-
 from py_markdown_table.markdown_table import markdown_table
+from scipy.spatial import Delaunay
 
-#-- LAZ datasets
-path_laz_2m  = "/Users/hugo/data/ahn4/04GN2_21.LAZ"
+import startinpy
+
+# -- LAZ datasets
+path_laz_2m = "/Users/hugo/data/ahn4/04GN2_21.LAZ"
 path_laz_33m = "/Users/hugo/data/ahn4/69EZ1_21.LAZ"
 
-#-- random datasets
+# -- random datasets
 rng = np.random.default_rng(seed=42)
 pts_2d_10k = rng.random((10000, 2))
 pts_2d_50k = rng.random((50000, 2))
@@ -26,243 +27,248 @@ pts_3d_50k = rng.random((50000, 3))
 
 t = []
 
+
 def t_delaunator():
     t.append({"library": "delaunator"})
-    #-- 10k
+    # -- 10k
     t1 = time.perf_counter()
     Delaunator(pts_2d_10k)
     t2 = time.perf_counter()
-    t[-1]["random_10k"] = "{:.3f}".format(t2-t1)
-    #-- 50k
+    t[-1]["random_10k"] = "{:.3f}".format(t2 - t1)
+    # -- 50k
     t1 = time.perf_counter()
     Delaunator(pts_2d_50k)
     t2 = time.perf_counter()
-    t[-1]["random_50k"] = "{:.3f}".format(t2-t1)
-    #-- LAZ_2M
+    t[-1]["random_50k"] = "{:.3f}".format(t2 - t1)
+    # -- LAZ_2M
     las = laspy.read(path_laz_2m)
     pts = np.vstack((las.x, las.y)).transpose()
     t1 = time.perf_counter()
     Delaunator(pts)
     t2 = time.perf_counter()
-    t[-1]["LAZ_2M"] = "{:.3f}".format(t2-t1)
-    #-- LAZ_33M
+    t[-1]["LAZ_2M"] = "{:.3f}".format(t2 - t1)
+    # -- LAZ_33M
     las = laspy.read(path_laz_33m)
     pts = np.vstack((las.x, las.y)).transpose()
     t1 = time.perf_counter()
     Delaunator(pts)
     t2 = time.perf_counter()
-    t[-1]["LAZ_33M"] = "{:.3f}".format(t2-t1)
+    t[-1]["LAZ_33M"] = "{:.3f}".format(t2 - t1)
     # t[-1]["LAZ_33M"] = "X"
-    #-- dem.tiff
-    d = rasterio.open('/Users/hugo/projects/startinpy/data/dem_01.tif')
+    # -- dem.tiff
+    d = rasterio.open("/Users/hugo/projects/startinpy/data/dem_01.tif")
     band1 = d.read(1)
     tr = d.transform
     pts = []
     for i in range(band1.shape[0]):
         for j in range(band1.shape[1]):
-             x = tr[2] + (j * tr[0]) + (tr[0] / 2)
-             y = tr[5] + (i * tr[4]) + (tr[4] / 2)
-             z = band1[i][j]
[…]
