### 2024-08-22 88e94b2
refactor: improvements to benchmarks
