+++ dt_comparisons/requirements.txt
@@ -1,0 +1,43 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements.in
+affine==2.4.0
+    # via rasterio
+attrs==24.2.0
+    # via rasterio
+certifi==2024.7.4
+    # via rasterio
+click==8.1.7
+    # via
+    #   click-plugins
+    #   cligj
+    #   rasterio
+click-plugins==1.1.1
+    # via rasterio
+cligj==0.7.2
+    # via rasterio
+laspy==2.5.4
+    # via -r requirements.in
+numpy==2.1.0
+    # via
+    #   -r requirements.in
+    #   laspy
+    #   rasterio
+    #   scipy
+    #   snuggs
+    #   triangle
+py-markdown-table==1.1.0
+    # via -r requirements.in
+pyparsing==3.1.2
+    # via snuggs
+rasterio==1.3.10
+    # via -r requirements.in
+scipy==1.14.1
+    # via -r requirements.in
+setuptools==73.0.1
+    # via rasterio
+snuggs==1.4.7
+    # via rasterio
+startinpy==0.10.2
+    # via -r requirements.in
+triangle==20230923
+    # via -r requirements.in
