+++ dt_comparisons/readme.md
@@ -1,11 +1,13 @@
+# Comparisons
 
-## Python library tested
+Rough benchmarks of various Delaunay triangulation libraries against random points, real-world point clouds, and a real world DEM.
 
-  1. [Delaunator](https://github.com/HakanSeven12/Delaunator-Python): pure Python port of a proven fast triangulator written original in JavaScript https://github.com/mapbox/delaunator. 
+## Python libraries tested
+
+  1. [Delaunator](https://github.com/HakanSeven12/Delaunator-Python): pure Python port of a proven fast triangulator written original in JavaScript <https://github.com/mapbox/delaunator>.
   2. [SciPy](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.Delaunay.html): wrapper around [Qhull](http://qhull.org/), written in C. Using the batch construction in 2D.
   3. [SciPy-inc](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.Delaunay.html): wrapper around [Qhull](http://qhull.org/), written in C. Using the incremental insertion (as startinpy does).
   4. [Triangle](https://pypi.org/project/triangle/): wrapper around the [fast and robust C library](https://www.cs.cmu.edu/~quake/triangle.html) that performs constrained DT and meshing
-
 
 ## Datasets
 
@@ -15,9 +17,7 @@
   4. __LAZ_33M__: a real-world subset of the [AHN4 dataset](https://www.ahn.nl/). The sub-tile [69EZ1_21.LAZ](https://geotiles.citg.tudelft.nl/AHN4_T/69EZ1_21.LAZ) contains 33,107,889 points.
   5. __dem.tiff__: the GeoTIFF file in `/data/` is a 550x505 gridded terrain. We take the centre of each cell, reading row-by-row and column-by-column, and this creates a 277,750 dataset of points that are collinear and cocircular with many others (degenerate cases for the DT).
 
-
 ## Results
-
 
 |            |random_10k|random_50k|LAZ_2M|LAZ_33M|dem.tiff|
 |:-----------|----------|---------:|-----:|------:|-------:|
@@ -29,19 +29,6 @@
 
 ## To replicate
 
-  1. install those packages:
-    
-    - numpy
-    - laspy
-    - rasterio
-    - time
-    - startinpy
-    - triangle
-    - scipy
-    - https://github.com/HakanSeven12/Delaunator-Python
-    - py_markdown_table
-
-  2. download the 2 LAZ files
-  3. change the path (lines 17+18)
-  4. `python comparisons.py`, this generates a summary table in Markdown
-
+  1. install those packages: `pip install -r requirements.txt`
+  2. download the 2 LAZ files and Delaunator.py: `./download`
+  3. `python comparisons.py`, this generates a summary table in Markdown
