+++ dt_comparisons/requirements.in
@@ -1,0 +1,7 @@
+laspy
+numpy
+py_markdown_table
+rasterio
+scipy
+startinpy
+triangle
