+++ dt_comparisons/requirements.in
@@ -1,5 +1,6 @@
 laspy
 numpy
+pdal
 py_markdown_table
 rasterio
 scipy
