### 2024-08-26 35f65a0
Update the dt_comparisons folder
add PDAL (based on the code of @gadomski , see #24)

Clean code and have instructions to download datasets

requirements.txt added
