+++ dt_comparisons/comparisons.py
@@ -2,20 +2,13 @@ import numpy as np
 import laspy
 import rasterio
 import time
-import startinpy
-import triangle 
-from scipy.spatial import Delaunay
-from delaunay.quadedge.mesh import Mesh
-from delaunay.quadedge.point import Vertex
-from delaunay.delaunay import delaunay
 from random import uniform
-from Delaunator import Delaunator
-
 from py_markdown_table.markdown_table import markdown_table
 
 #-- LAZ datasets
-path_laz_2m  = "/Users/hugo/data/ahn4/04GN2_21.LAZ"
-path_laz_33m = "/Users/hugo/data/ahn4/69EZ1_21.LAZ"
+path_laz_2m  = "./04GN2_21.LAZ"
+path_laz_33m = "./69EZ1_21.LAZ"
+path_dem_01 = "../data/dem_01.tif"
 
 #-- random datasets
 rng = np.random.default_rng(seed=42)
@@ -26,8 +19,65 @@ pts_3d_50k = rng.random((50000, 3))
 
 t = []
 
+def t_pdal():
+    from pdal import Filter, Reader
+
+    def make_pdal_points(points):
+        return np.array(
+            [(x, y, z) for x, y, z in zip(points[:, 0], points[:, 1], points[:, 2])],
+            dtype=[("X", float), ("Y", float), ("Z", float)],
+        )
+    t.append({"library": "Delaunator-pdal"})
+    #-- 10k
+    points = make_pdal_points(pts_3d_10k)
+    t1 = time.perf_counter()
+    Filter.delaunay().pipeline(points).execute()
+    t2 = time.perf_counter()
+    t[-1]["random_10k"] = "{:.3f}".format(t2 - t1)
+    #-- 50k
+    points = make_pdal_points(pts_3d_50k)
+    t1 = time.perf_counter()
+    Filter.delaunay().pipeline(points).execute()
+    t2 = time.perf_counter()
+    t[-1]["random_50k"] = "{:.3f}".format(t2 - t1)
+    #-- dem.tiff
+    d = rasterio.open(path_dem_01)
+    band1 = d.read(1)
+    tr = d.transform
+    pts = []
+    for i in range(band1.shape[0]):
+        for j in range(band1.shape[1]):
+            x = tr[2] + (j * tr[0]) + (tr[0] / 2)
+            y = tr[5] + (i * tr[4]) + (tr[4] / 2)
+            z = band1[i][j]
+            if z != d.nodatavals:
+                pts.append([x, y, z])
+    points = make_pdal_points(np.array(pts))
+    t1 = time.perf_counter()
+    Filter.delaunay().pipeline(points).execute()
+    t2 = time.perf_counter()
+    t[-1]["dem.tiff"] = "{:.3f}".format(t2 - t1)
+    #-- LAZ_2M
+    pipeline = Reader.las(path_laz_2m).pipeline()
+    pipeline.execute()
+    t1 = time.perf_counter()
+    Filter.delaunay().pipeline(pipeline.arrays[0]).execute()
+    t2 = time.perf_counter()
+    t[-1]["LAZ_2M"] = "{:.3f}".format(t2 - t1)
+    #-- LAZ_33M
+    pipeline = Reader.las(path_laz_33m).pipeline()
+    pipeline.execute()
+    t1 = time.perf_counter()
+    Filter.delaunay().pipeline(pipeline.arrays[0]).execute()
+    t2 = time.perf_counter()
+    t[-1]["dem.tiff"] = "{:.3f}".format(t2-t1) 
+    t[-1]["LAZ_33M"] = "{:.3f}".format(t2 - t1)
+
+
+
 def t_delaunator():
-    t.append({"library": "delaunator"})
+    from Delaunator import Delaunator
[…]
