+++ dt_comparisons/requirements.txt
@@ -1,0 +1,16 @@
+affine==2.4.0
+attrs==24.2.0
+certifi==2024.7.4
+click==8.1.7
+click-plugins==1.1.1
+cligj==0.7.2
+laspy==2.5.4
+numpy==2.1.0
+pdal==3.4.5
+py-markdown-table==1.1.0
+pyparsing==3.1.4
+rasterio==1.3.10
+scipy==1.14.1
+snuggs==1.4.7
+startinpy==0.11.0
+triangle2==20230923
