+++ dt_comparisons/requirements.in
@@ -1,0 +1,8 @@
+laspy
+numpy
+pdal
+py_markdown_table
+rasterio
+scipy
+startinpy
+triangle
