### 2024-08-27 b33d1cb
Add pdal comparison to the docs comparison
