+++ docs/comparison.md
@@ -1,22 +1,23 @@
 # Comparison 
 
-The two tables below compare startinpy to its main alternatives.
+The two tables below compare startinpy to a few of its main alternatives:
 
-The [Delaunator package](https://github.com/HakanSeven12/Delaunator-Python) is a pure Python port of a proven fast triangulator [originally written in JavaScript](https://github.com/mapbox/delaunator).
-[SciPy](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.Delaunay.html) is `scipy.spatial.Delaunay`, and SciPy-inc is the variation where an incremental algorithm is used (instead of a batch process).
-[Triangle](https://pypi.org/project/triangle/) is the Python bindings of the C code.
+* [__Delaunator-py__](https://github.com/HakanSeven12/Delaunator-Python) is a pure Python port of a proven fast triangulator [originally written in JavaScript](https://github.com/mapbox/delaunator).
+* [__Delaunay-pdal__](https://pdal.io/en/2.7.2/stages/filters.delaunay.html) is the Delaunay triangulator of [PDAL](https://pdal.io), a general point data library that reads, filters, and writes point clouds. It uses a [CPP implementation of the Delaunator algorithm](https://github.com/delfrrr/delaunator-cpp).
+* [__SciPy__](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.Delaunay.html) is `scipy.spatial.Delaunay`, and __SciPy-inc__ is the variation where an incremental algorithm is used (instead of a batch process).
+* [__Triangle__](https://pypi.org/project/triangle/) is the Python bindings of the C code.
 
-|                        | Delaunator | SciPy | SciPy-inc | Triangle | startinpy |
-|------------------------|:----------:|:-----:|:---------:|:--------:|:---------:|
-| constrained DT         |    --      | --    |     --    |   +      |   --      |
-| data structure exposed |    +       | --    |     --    |   --     |   +       |
-| efficient searches     |    +       | +     |     +     |   --     |   +       |
-| elevation/2.5D         |    --      | --    |     --    |   --     |   +       |
-| extra attributes       |    --      | --    |     --    |   --     |   +       |
-| speed                  |    --      | --    |     --    |   ++     |   +       |
-| incremental insertion  |    --      | --    |     +     |   --     |   +       |
-| remove vertices        |    --      | --    |     --    |   --     |   +       |
-| xy-duplicate handling  |    --      | --    |     --    |   --     |   +       |
+|                        | Delaunator-py | Delaunator-pdal | SciPy | SciPy-inc | Triangle | startinpy |
+|------------------------|:-------------:|:---------------:|:-----:|:---------:|:--------:|:---------:|
+| constrained DT         |    --         |    --           | --    |     --    |   +      |   --      |
+| data structure exposed |    +          |    --           | --    |     --    |   --     |   +       |
+| efficient searches     |    +          |    --           | +     |     +     |   --     |   +       |
+| elevation/2.5D         |    --         |    +            | --    |     --    |   --     |   +       |
+| extra attributes       |    --         |    --           | --    |     --    |   --     |   +       |
+| speed                  |    --         |    ++           | --    |     --    |   ++     |   +       |
+| incremental insertion  |    --         |    --           | --    |     +     |   --     |   +       |
+| remove vertices        |    --         |    --           | --    |     --    |   --     |   +       |
[…]
