### 2024-08-27 f1d9efc
Update the DT speed comparisons numbers
I ran again the whole tests
