### 2024-09-15 hugoledoux/startinpy#26 @weiji14
AttributeError: 'builtins.DT' object has no attribute 'add_attribute_map'
When running the example at https://startinpy.readthedocs.io/0.11.0/attributes.html#retrieving-the-extra-attributes:

```python
dt = startinpy.DT()
dt.add_attribute_map(np.dtype([("classification", np.uint32)]))
dt.insert_one_pt([85000.0, 444003.2, 2.2], classification=6)
```

The following AttributeError is raised:

```python-traceback
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: 'builtins.DT' object has no attribute 'add_attribute_map'
```

Has the `add_attribute_map` method been renamed to something else? Documentation should be updated to reflect the new method.

Xref review at https://github.com/openjournals/joss-reviews/issues/7123
