### 2024-09-15 #7123 @weiji14
Just quickly mentioning that I've done an initial round of reviewing the paper and repository code (https://github.com/openjournals/joss-reviews/issues/7123#issuecomment-2351796864), and have been pleased so far with the quality of the software and writing! So far, I have only two minor suggestions (reported at https://github.com/hugoledoux/startinpy/issues/25 and https://github.com/hugoledoux/startinpy/issues/26) for @hugoledoux to resolve, that can hopefully be addressed with a couple of small PRs/commits.

I do have two remaining items on my checklist in terms of verifying the Performance benchmarks and a more thorough Reproducibility check that will both take more time, but will try to get those done this or next week.
