### 2024-09-16 339ab8e
Add NumPY as required dependency
fixes #25

Removed the `requires-dist` since the better/improved way seems to be
with `[dependencies]`, see https://www.maturin.rs/metadata
