+++ pyproject.toml
@@ -5,10 +5,14 @@ build-backend = "maturin"
 [project]
 name = "startinpy"
 requires-python = ">=3.9"
-requires-dist = ["toml>=0.10.2"]
 classifiers = [
     "Programming Language :: Rust",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
+]
+
+dependencies = [
+    "toml>=0.10.2", 
+    "numpy>=2.0.0"
 ]
 
