### 2024-09-16 55b7656
Fix name function in docs
the name changed and forgot that one

fixes #26
