### 2024-09-16 5c8ef29
Remove support for int32, uint32 for the attributes schema
Rust serde Value are 64, and I was accepting only those and casting
to 64bits in the background so better to just say that all attributes
are 64bits (except bool and String)
