### 2024-09-16 hugoledoux/startinpy#26 @hugoledoux
Indeed, the name of that function changed and I forgot that snippet. Thanks for catching that one.

Also, while going through the schema functions, I noticed that np.int32 and np.uint32 don't work because I use Rust serde Value and 64bits numbers are default, so I removed the 32bits (signed) integers for the schema, see 5c8ef298b89c78c39b7c54d8518a2a14cd9416aa
