### 2024-09-20 6727173
Add lazrs to dt_comparisons/requirements.txt
Fix `laspy.errors.LaspyException: No LazBackend selected, cannot decompress data` when running `laspy.read()`.
