### 2024-09-20 hugoledoux/startinpy#26 @weiji14
Missed one more `add_attribute_map` -> `set_attributes_schema` rename here:

https://github.com/hugoledoux/startinpy/blob/5c8ef298b89c78c39b7c54d8518a2a14cd9416aa/src/lib.rs#L434
