### 2024-09-20 #7123 @weiji14
Ok, I've completed running the performance benchmarks and am satisfied that the results are similar to those reported in the [comparison table](https://github.com/hugoledoux/startinpy/blob/f1d9efc8e6353fc4d5c4adfc0610d6d197604e4a/dt_comparisons/readme.md#results). I've did some quick experiments with using [`cargo-wizard`](https://github.com/kobzol/cargo-wizard) to see if I can optimize the Rust compilation and get more runtime performance for `startinpy`, but the numbers didn't seem to move much, so short of re-writing the algorithms with fancy SIMD operators, I'm happy with the current state of the library.

I also see that @hugoledoux has resolved my suggestions above, and so I'm happy with accepting this submission from my side. Well done!
