+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name    = "startinpy"
-version = "0.11.0"
+version = "0.11.1-beta1"
 authors = ["Hugo Ledoux <h.ledoux@tudelft.nl>"]
 license = "MIT"
 homepage="https://github.com/hugoledoux/startinpy"
@@ -11,7 +11,7 @@ edition = "2021"
 
 
 [dependencies]
-startin = "0.8.0"
+startin = "0.8.1"
 # startin = { path = "../startin" }
 # startin = { git = "https://github.com/hugoledoux/startin", branch = "develop" }
 numpy = "0.19"
