### 2024-10-07 a23720c
Add Lawson72 reference + reference the one having a DOI
fixes #28
