+++ paper/ref.bib
@@ -18,16 +18,18 @@
   volume = {64},
   year = {2009}}
 
-@techreport{Lawson72,
-  address = {Pasadena, CA, USA},
-  annote = {Same stuff as Lawson77: incremental construction by flips of 2D DT.},
+
+@article{Lawson72-1,
   author = {Lawson, Charles L.},
-  institution = {Jet Propulsion Laboratory},
-  keywords = {construction of VD/DT},
-  number = {Memo 299},
-  publisher = {Jet Propulsion Laboratory},
-  title = {{G}eneration of a triangular grid with applications to contour plotting},
+  date-modified = {2024-10-07 08:03:49 +0200},
+  doi = {10.1016/0012-365X(72)90093-3},
+  journal = {Discrete Applied Mathematics},
+  keywords = {bistellar flips},
+  pages = {365--372},
+  title = {{T}ransforming triangulations},
+  volume = {3},
   year = {1972}}
+
 
 
 @mastersthesis{Kan24,
