### 2024-10-07 cd7d7c1
Improve the paper with 2 small modifications
