### 2024-10-22 #7123 @hugoledoux
great, thanks everyone, that was useful!

I'm off later this week so that will be beginning of next week. And I was planning to release v1.0 (what better occasion than a JOSS paper to do that!) based on your feedback, so I need a few hours to think about all this...
