### 2024-10-23 8c2e10d
Remove the info in the `dt_comparison` folder and just point to the docs
