### 2024-10-23 3acf9da
Fix wrong syntax in a unit test
