### 2024-10-23 71d1519
Ruff the python files in the repo
from now on Ruff is used thus
