+++ tests/test_spatialextent.py
@@ -1,29 +1,35 @@
+import numpy as np
 import pytest
 import startinpy
-import numpy as np
+
 
 def dt_5_points():
     dt = startinpy.DT()
-    pts = np.array([
-        [0.0, 0.0, 1.0],
-        [10.0, 0.0, 2.0],
-        [10.0, 10.0, 3.0],
-        [0.0, 10.0, 4.0],
-        [5.0, 5.0, 5.0]
-     ])
+    pts = np.array(
+        [
+            [0.0, 0.0, 1.0],
+            [10.0, 0.0, 2.0],
+            [10.0, 10.0, 3.0],
+            [0.0, 10.0, 4.0],
+            [5.0, 5.0, 5.0],
+        ]
+    )
     dt.insert(pts)
     return dt
+
 
 def random(n=20):
     rng = np.random.default_rng()
     pts = rng.random((n, 3))
     pts = pts * 100
-    return pts 
+    return pts
+
 
 def test_bbox_empty():
     dt = startinpy.DT()
     bbox = dt.get_bbox()
     assert np.isinf(bbox).all() == True
+
 
 def test_bbox():
     pts = random(100)
@@ -34,6 +40,7 @@ def test_bbox():
     assert bbox[2] > bbox[0]
     assert bbox[3] > bbox[1]
 
+
 def test_bbox_square():
     dt = dt_5_points()
     bbox = dt.get_bbox()
@@ -42,19 +49,21 @@ def test_bbox_square():
     assert bbox[2] == pytest.approx(10.0)
     assert bbox[3] == pytest.approx(10.0)
 
+
 def test_convexhull_empty():
     dt = startinpy.DT()
     ch = dt.convex_hull()
     assert len(ch) == 0
-    dt.insert_one_pt([1., 1., 1.])
+    dt.insert_one_pt([1.0, 1.0, 1.0])
     ch = dt.convex_hull()
     assert len(ch) == 0
-    dt.insert_one_pt([2., 3., 4.])
+    dt.insert_one_pt([2.0, 3.0, 4.0])
     ch = dt.convex_hull()
     assert len(ch) == 0
-    dt.insert_one_pt([1., 6., 4.])
+    dt.insert_one_pt([1.0, 6.0, 4.0])
     ch = dt.convex_hull()
     assert len(ch) == 3
+
 
 def test_convexhull():
     dt = dt_5_points()
@@ -62,20 +71,21 @@ def test_convexhull():
     assert len(ch) == 4
     assert dt.is_vertex_convex_hull(1) == True
     with pytest.raises(OverflowError):
-        dt.is_vertex_convex_hull(-1) 
+        dt.is_vertex_convex_hull(-1)
     assert dt.is_vertex_convex_hull(5) == False
     assert dt.is_inside_convex_hull([7.1, 2.1]) == True
     assert dt.is_inside_convex_hull([-7.1, 2.1]) == False
+
 
 def test_locate():
     dt = dt_5_points()
     assert (dt.locate([7.1, 2.1]) == np.array([5, 1, 2])).all()
     with pytest.raises(Exception):
-        dt.locate(-1., 9.0)   
+        dt.locate(-1.0, 9.0)
+
 
 def test_closest_point():
     dt = dt_5_points()
     assert dt.closest_point([7.1, 2.1]) == 5
     with pytest.raises(Exception):
-        dt.locate(-1., 9.0)           
-    
+        dt.locate(-1.0, 9.0)
