+++ tests/test_init.py
@@ -1,6 +1,6 @@
+import numpy as np
 import pytest
 import startinpy
-import numpy as np
 
 
 def test_init():
@@ -8,44 +8,49 @@ def test_init():
     assert dt.number_of_vertices() == 0
     assert dt.number_of_triangles() == 0
     assert dt.get_attributes_schema() == []
-    dt = startinpy.DT(np.dtype([('classification', np.float32), ('visited', bool)]))
+    dt = startinpy.DT(np.dtype([("classification", np.float32), ("visited", bool)]))
     assert dt.number_of_vertices() == 0
     assert dt.number_of_triangles() == 0
-    assert dt.get_attributes_schema() == np.dtype([('classification', np.float32), ('visited', bool)])
+    assert dt.get_attributes_schema() == np.dtype(
+        [("classification", np.float32), ("visited", bool)]
+    )
+
 
 def test_wrong_array_size():
     dt = startinpy.DT()
     with pytest.raises(Exception):
-        dt.insert_one_pt([0., 0.])
+        dt.insert_one_pt([0.0, 0.0])
     with pytest.raises(Exception):
-        dt.insert_one_pt([0., 0., 1., 2.])
+        dt.insert_one_pt([0.0, 0.0, 1.0, 2.0])
 
 
 def test_before_first_triangle():
     dt = startinpy.DT()
-    dt.insert_one_pt([0., 0., 0.])
+    dt.insert_one_pt([0.0, 0.0, 0.0])
     assert dt.number_of_vertices() == 1
     assert dt.number_of_triangles() == 0
-    dt.insert_one_pt([1., 0., 0.])
+    dt.insert_one_pt([1.0, 0.0, 0.0])
     assert dt.number_of_vertices() == 2
     assert dt.number_of_triangles() == 0
-    dt.insert_one_pt([1., 1., 0.])
+    dt.insert_one_pt([1.0, 1.0, 0.0])
     assert dt.number_of_vertices() == 3
     assert dt.number_of_triangles() == 1
 
+
 def test_init_phase_duplicates_remove():
     dt = startinpy.DT()
-    dt.insert_one_pt([0., 0., 0.])
-    dt.insert_one_pt([1., 0., 0.])
-    dt.insert_one_pt([1., 1., 0.])
+    dt.insert_one_pt([0.0, 0.0, 0.0])
+    dt.insert_one_pt([1.0, 0.0, 0.0])
+    dt.insert_one_pt([1.0, 1.0, 0.0])
     assert dt.number_of_vertices() == 3
     assert dt.number_of_triangles() == 1
-    dt.insert_one_pt([1., 0., 0.])
+    dt.insert_one_pt([1.0, 0.0, 0.0])
     assert dt.number_of_vertices() == 3
     assert dt.number_of_triangles() == 1
     dt.remove(3)
     assert dt.number_of_vertices() == 2
-    assert dt.number_of_triangles() == 0        
+    assert dt.number_of_triangles() == 0
+
 
 def test_grid():
     dt = startinpy.DT()
@@ -54,14 +59,15 @@ def test_grid():
             dt.insert_one_pt([float(i), float(j), 1.0])
     assert dt.number_of_vertices() == 100
 
+
 def test_collinear():
     dt = startinpy.DT()
-    dt.insert_one_pt([0., 0., 0.])
-    dt.insert_one_pt([1., 0., 0.])
-    dt.insert_one_pt([2., 0., 0.])
+    dt.insert_one_pt([0.0, 0.0, 0.0])
+    dt.insert_one_pt([1.0, 0.0, 0.0])
+    dt.insert_one_pt([2.0, 0.0, 0.0])
[…]
