### 2024-10-23 483f26e
Replace snippets in docs by inserting the python files
