+++ docs/attributes.md
@@ -75,31 +75,6 @@ dt.attributes[1:]['classification']
 
 ## One full Python example
 
-```python
-import startinpy
-import numpy as np
-import json
-import laspy
-
-las = laspy.read("../data/small.laz")
-#-- read intensity and store it as extra_attribute in the startinpy DT
-d = np.vstack((las.x, las.y, las.z, las.classification)).transpose()
-d = d[::1] #-- thinning to speed up, put ::1 to keep all the points
-
-print(d)
-
-dt = startinpy.DT(np.dtype([("classification", np.uint64)]))
-for each in d:
-    dt.insert_one_pt(each[:3], classification=int(each[3]))
-
-dt.set_vertex_attributes(50, classification=int(112.2))
-
-print(dt)
-print(dt.get_attributes_schema())
-print(dt.attributes[1:]['classification'])
-
-a = dt.get_vertex_attributes(50)
-print("=>", a)
-a = dt.get_vertex_attributes(49)
-print("=>", a)
+```{literalinclude} ../demo/attributes.py
 ```
+
