+++ docs/examples.md
@@ -4,49 +4,19 @@
 
 We recommend you use [laspy](https://laspy.readthedocs.io) (`pip install 'laspy[lazrs]'` to install) to read a LAS/LAZ into a NumPy array, and then pass that array to startinpy:
 
-```python
-import startinpy
-import numpy as np
-import laspy
-
-las = laspy.read("myfile.laz")
-pts = np.vstack((las.x, las.y, las.z)).transpose()
-pts = pts[::1] #-- thinning to speed up, put ::10 to keep 1/10 of the points
-dt = startinpy.DT()
-dt.insert(pts)
-print("number vertices:", dt.number_of_vertices())
+```{literalinclude} ../demo/example_reading_laz.py
 ```
+
 
 ## Exporting the DT to GeoJSON
 
-```python
-import startinpy
-import numpy as np
-
-#-- generate 100 points randomly in the plane
-rng = np.random.default_rng(seed=42)
-pts = rng.random((100, 3))
-dt = startinpy.DT()
-dt.insert(pts, insertionstrategy="AsIs")
-dt.write_geojson("myfile.geojson")
+```{literalinclude} ../demo/example_exporting_geojson.py
 ```
+
 
 ## Exporting the DT to several mesh formats with [meshio](https://github.com/nschloe/meshio)
 
-```python
-import startinpy
-import meshio
-import laspy
-import numpy as np
-
-las = laspy.read("../data/small.laz")
-pts = np.vstack((las.x, las.y, las.z)).transpose()
-dt = startinpy.DT()
-dt.insert(pts)
-vs = dt.points
-vs[0] = vs[1] #-- to ensure that infinite vertex is not blocking the viz
-cells = [("triangle", dt.triangles)]
-meshio.write_points_cells("mydt.vtu", vs, cells)
+```{literalinclude} ../demo/example_exporting_meshio.py
 ```
 
 ## Reading a GeoTIFF file with rasterio
@@ -61,28 +31,7 @@ The `no_data` values are not inserted in the triangulation.
 
 This code saves the resulting triangulation to a [PLY file](<https://en.wikipedia.org/wiki/PLY_(file_format)>) that can be opened directly in QGIS (with the newish [MDAL mesh](https://docs.qgis.org/3.34/en/docs/user_manual/working_with_mesh/mesh_properties.html)).
 
-```python
-import startinpy
-import rasterio
-import random
-
-d = rasterio.open('../data/dem_01.tif')
-band1 = d.read(1)
-t = d.transform
-pts = []
-for i in range(band1.shape[0]):
-    for j in range(band1.shape[1]):
-         x = t[2] + (j * t[0]) + (t[0] / 2)
-         y = t[5] + (i * t[4]) + (t[4] / 2)
-         z = band1[i][j]
-         #-- skip no_data + select randomly only 1% of the points
-         if (z != d.nodatavals) and (random.randint(0, 100) == 5):
-             pts.append([x, y, z])
-dt = startinpy.DT()
-dt.insert(pts, insertionstrategy="BBox")
-#-- exaggerate the elevation by a factor 2.0
-dt.vertical_exaggeration(2.0)
-dt.write_ply("mydt.ply")
+```{literalinclude} ../demo/example_reading_geotiff.py
 ```
 
 ```{image} figs/mdal.jpg
@@ -94,34 +43,7 @@ dt.write_ply("mydt.ply")
 
 You need to install [Polyscope](https://polyscope.run/py/) (basically `pip install polyscope`).
 
-```python
-import startinpy
-import numpy as np
-import polyscope as ps
-import laspy
-
-las = laspy.read("data/small.laz")
-pts = np.vstack((las.x, las.y, las.z)).transpose()
[…]
