### 2024-10-23 5f979d2
Add comments to showcase and copy to the readme
