+++ changelog.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2024-07-17
+### Changed
+- interpolation with NNI can now use the "precompute" of Voronoi vertices (speed up in many cases) 
+- bindings now compiled for Python 3.12 with github actions and for linux
+- many changes the docs
+- upgrade to latest startin (which fixed a few bugs)
+### Added
+- comparison of DT speed with alternatives added
+- folder `./paper/` contains the JOSS paper that was submitted and reviewed and accepted
+
 ## [0.11.0] - 2024-07-17
 ### Changed
 - breaking change: all functions taking one location as argument (in 2D or 3D), now take an array `[x, y, z]` instead of the separate coordinates as different arguments. This harmonise all functions, and is better when dealing with NumPy. Functions affected: `insert_one_pt()`, `closest_point()`, `locate()`, and `is_inside_convex_hull()`.
@@ -83,6 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new name "startinpy" to avoid clashes with the rust lib "startin" on which this project is based on. Since I use a Rust builder (maturin), there were clashes and probably a better idea to call it "startinpy" anyway.
 
 
+[0.12.0]: https://github.com/hugoledoux/startin/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/hugoledoux/startin/compare/0.10.2...0.11.0
 [0.10.2]: https://github.com/hugoledoux/startin/compare/0.10.1...0.10.2
 [0.10.1]: https://github.com/hugoledoux/startin/compare/0.10.0...0.10.1
