+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name    = "startinpy"
-version = "0.11.1-beta1"
+version = "0.12.0"
 authors = ["Hugo Ledoux <h.ledoux@tudelft.nl>"]
 license = "MIT"
 homepage="https://github.com/hugoledoux/startinpy"
