+++ changelog.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.0] - 2024-07-17
+## [0.12.0] - 2024-10-28
 ### Changed
 - interpolation with NNI can now use the "precompute" of Voronoi vertices (speed up in many cases) 
 - bindings now compiled for Python 3.12 with github actions and for linux
