### 2024-10-28 267d632
Fix wrong ULR in changelog
