### 2024-10-28 90a53c2
Fix errors in publish.yml because of multi-cursors...
