### 2024-10-28 #7123 @hugoledoux
@mikemahoney218

==> 0.12.0 (https://github.com/hugoledoux/startinpy/releases/tag/0.12.0)
==> https://doi.org/10.5281/zenodo.14001126
