### 2024-11-11 #7123 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
