### 2024-11-11 #7123 @kthyng
@hugoledoux
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. ("delaunay" is not capitalized, but please check other words too)
