+++ paper/ref.bib
@@ -3,7 +3,7 @@
   author = {Aurenhammer, Franz and Klein, Rolf and Lee, Der-Tsai},
   doi = {10.1142/8685},
   publisher = {World Scientific},
-  title = {Voronoi Diagrams and Delaunay Triangulations},
+  title = {Voronoi diagrams and {D}elaunay triangulations},
   year = {2012}}
 
 
@@ -14,7 +14,7 @@
   doi = {10.1016/j.isprsjprs.2008.09.007},
   number = {1},
   pages = {1--16},
-  title = {Full-waveform topographic lidar: State-of-the-art},
+  title = {Full-waveform topographic lidar: {S}tate-of-the-art},
   volume = {64},
   year = {2009}}
 
