### 2024-11-12 19a9654
Fix the weird typo in the CI yaml file
