+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Ledoux
+  given-names: Hugo
+  orcid: "https://orcid.org/0000-0002-1251-8654"
+doi: 10.5281/zenodo.14001126
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Ledoux
+    given-names: Hugo
+    orcid: "https://orcid.org/0000-0002-1251-8654"
+  date-published: 2024-11-12
+  doi: 10.21105/joss.07123
+  issn: 2475-9066
+  issue: 103
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 7123
+  title: "startinpy: A Python library for modelling and processing 2.5D
+    triangulated terrains"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.07123"
+  volume: 9
+title: "startinpy: A Python library for modelling and processing 2.5D
+  triangulated terrains"
