### 2024-11-12 #7123 @hugoledoux
thanks everyone, the package is better because of you and the JOSS process
