### 2024-11-12 #7123 @kthyng
Congratulations on your new publication @hugoledoux! Many thanks to editor @mikemahoney218 and to reviewers @weiji14,  @kylemann16, and @gadomski for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
